backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-filter==24.2
drf-spectacular==0.27.2
whitenoise==6.6.0
gunicorn==22.0.0
